BUGFIX: Atmosphere is now correctly generated.
We were assuming that in Planet.GeneratePlanet that comes from StarFormNET, when SurfacePressure was used it meant that the Atmopshere was completely created. However it is not the case since only the surfpres was created and then the full Atmopshere completed way after. SurfacePressure is now treated as a whole for some calculation and then the Atmosphere is created.